Создал папки и разложил логику ядро и UI
