+++ FileManager/FileManager/FileManager.csproj
@@ -9,4 +9,8 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="ConsoleUI\" />
+  </ItemGroup>
+
 </Project>
