+++ FileManager/FileManager/Logic/BasicLogic.cs
