+++ FileManager/FileManager/Logic/DirectiveCLS.cs
