+++ FileManager/FileManager/Logic/DirectiveDIR.cs
