+++ FileManager/FileManager/Logic/DirectiveEXIT.cs
@@ -14,7 +14,7 @@ namespace FileManager
 
         public void RunDirective(params string[] args)
         {
-            StartDirective(args[0]);
+            ExitCommandExecuter();
         }
