+++ FileManager/FileManager/Logic/DirectiveHELP.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Text;
 
 namespace FileManager
