+++ FileManager/FileManager/Logic/DirectiveMD.cs
