+++ FileManager/FileManager/Logic/DirectiveRD.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Text;
 
 namespace FileManager
 {
-    internal class DirectiveMOVE : IDirective
+    internal class DirectiveRD : IDirective
     {
         private const string _directiveName = "MOVE";
         public string DirectiveName { get => _directiveName; }
@@ -14,20 +15,23 @@ namespace FileManager
             throw new NotImplementedException();
         }
 
+
         /// <summary>
-        /// Метод перемещает каталог со всем содержимым или файл
+        /// Метод перемещает директорию
         /// </summary>
-        /// <param name="pathFrom">string Путь откуда перемешать</param>
-        /// <param name="pathTo">string Путь куда перемешать</param>
-        public static void MoveCommandExecuter(string pathFrom, string pathTo)
+        /// <param name="pathFrom">string Путь откуда перемещать</param>
+        /// <param name="pathTo">string Путь куда перемещать</param>
+        public static void RemoveDirectoryCommandExecuter()
         {
             if (UserCommands.Command.Split().Length == 3)
             {
-                if ((pathFrom.Contains(":\\") || pathFrom.Contains(":/"))
-                    && (pathTo.Contains(":\\") || pathTo.Contains(":/")))
-                {
-                    FileInfo file = new FileInfo(pathFrom);
+                string pathFrom = UserCommands.Command.Split()[1];
 
+                string pathTo = UserCommands.Command.Split()[2];
+
+                if ((pathFrom.Contains(":\\") || pathFrom.Contains(":/")) && !pathFrom.Contains('.')
+                    && !pathTo.Contains('.') && (pathTo.Contains(":\\") || pathTo.Contains(":/")))
+                {
                     DirectoryInfo directory = new DirectoryInfo(pathFrom);
 
                     try
@@ -35,10 +39,6 @@ namespace FileManager
                         if (directory.Exists)
                         {
                             directory.MoveTo(pathTo);
-                        }
-                        else if (file.Exists)
-                        {
-                            file.MoveTo(pathTo);
                         }
                     }
                     catch (Exception ex)
@@ -52,10 +52,6 @@ namespace FileManager
                 }
             }
         }
-
-
-
-
