+++ FileManager/FileManager/Repos/Exeptions.cs
